Added lots of error checking, replaced all '        ' with tabs

+++ vexim/adminfail.php
@@ -20,12 +20,14 @@
       <?
 	$query = "SELECT user_id,localpart FROM users WHERE domain_id='{$_COOKIE['vexim'][2]}' AND users.type='fail' ORDER BY localpart;";
 	$result = $db->query($query);
-	while ($row = $result->fetchRow()) {
-	  print "<tr>";
-	  print "<td align=\"center\"><a href=\"adminfaildelete.php?user_id={$row['user_id']}\">";
-	  print "<img style='border:0;width:10px;height:16px' src=\"images/trashcan.gif\" title=\"Delete fail {$row['localpart']}\"></a></td>\n";
-	  print "<td><a href=\"adminfailchange.php?user_id={$row['user_id']}\">{$row['localpart']}@{$_COOKIE['vexim'][1]}</a></td>\n";
-	  print "</tr>\n";
+	if ($result->numRows()) {
+	  while ($row = $result->fetchRow()) {
+	    print "<tr>";
+	    print "<td align=\"center\"><a href=\"adminfaildelete.php?user_id={$row['user_id']}\">";
+	    print "<img style='border:0;width:10px;height:16px' src=\"images/trashcan.gif\" title=\"Delete fail {$row['localpart']}\"></a></td>\n";
+	    print "<td><a href=\"adminfailchange.php?user_id={$row['user_id']}\">{$row['localpart']}@{$_COOKIE['vexim'][1]}</a></td>\n";
+	    print "</tr>\n";
+	  }
 	}
       ?>
     </table>
