+++ vexim/site.php
@@ -69,32 +69,36 @@
 			WHERE users.domain_id = domains.domain_id
 			AND domain !='admin' AND admin=1 ORDER BY domain";
   	$result = $db->query($query);
-	while ($row = $result->fetchRow()) {
-	  print "<tr>";
-	  print "\t<td><a href=\"sitedelete.php?domain_id={$row['domain_id']}\">";
-	  print "<img style='border:0;width:10px;height:16px' title='Delete {$row['domain']}' src='images/trashcan.gif' alt='trashcan'></a></td>\n";
-	  print "\t<td><a href=\"sitechange.php?domain_id={$row['domain_id']}\">{$row['domain']}</a></td>\n";
-	  print "\t<td>{$row['localpart']}@{$row['domain']}</td>\n";
-	  print "</tr>\n";
+	if ($result->numRows()) {
+	  while ($row = $result->fetchRow()) {
+	    print "<tr>";
+	    print "\t<td><a href=\"sitedelete.php?domain_id={$row['domain_id']}\">";
+	    print "<img style='border:0;width:10px;height:16px' title='Delete {$row['domain']}' src='images/trashcan.gif' alt='trashcan'></a></td>\n";
+	    print "\t<td><a href=\"sitechange.php?domain_id={$row['domain_id']}\">{$row['domain']}</a></td>\n";
+	    print "\t<td>{$row['localpart']}@{$row['domain']}</td>\n";
+	    print "</tr>\n";
+	  }
 	}
       ?>
       <tr><td></td></tr>
       <tr><td colspan="3"><b>WARNING:</b> Deleting a domain will delete all user accounts in that domain permanently!</td></tr>
       <tr><td></td></tr>
       <tr>
-        <th></th>
+	<th></th>
 	<th>Relay domains</th>
       </tr>
       <?
-        $query = "SELECT domain,domain_id FROM domains WHERE domain !='admin' AND type='relay' ORDER BY domain";
-        $result = $db->query($query);
-        while ($row = $result->fetchRow()) {
-          print "<tr>";
-          print "\t<td><a href=\"sitedelete.php?domain_id={$row['domain_id']}\">";
-          print "<img style='border:0;width:10px;height:16px' title='Delete{$row['domain']}' src='images/trashcan.gif' alt='trashcan'></a></td>\n";
-          print "\t<td>{$row['domain']}</a></td>\n";
-          print "</tr>\n";
-        }
+	$query = "SELECT domain,domain_id FROM domains WHERE domain !='admin' AND type='relay' ORDER BY domain";
+	$result = $db->query($query);
+	if ($result->numRows()) {
+	  while ($row = $result->fetchRow()) {
+	    print "<tr>";
+	    print "\t<td><a href=\"sitedelete.php?domain_id={$row['domain_id']}\">";
+	    print "<img style='border:0;width:10px;height:16px' title='Delete{$row['domain']}' src='images/trashcan.gif' alt='trashcan'></a></td>\n";
+	    print "\t<td>{$row['domain']}</a></td>\n";
+	    print "</tr>\n";
+	  }
+	}
       ?>
     </table>
     </div>
